+++ src/app/data/peopleData.ts
@@ -1,0 +1,26 @@
+export const peopleData = [
+    {
+      name: 'Alice Johnson',
+      role: 'AI Developer',
+      description:
+        'Alice uses the Explorer to ensure her AI models align with ethical guidelines and organizational values, testing for bias and safety in deployment.',
+    },
+    {
+      name: 'Bob Patel',
+      role: 'Policy Analyst',
+      description:
+        'Bob leverages the tool to evaluate AI responses against regulatory standards, ensuring compliance with public safety and fairness policies.',
+    },
+    {
+      name: 'Clara Nguyen',
+      role: 'Ethics Researcher',
+      description:
+        'Clara studies AI behavior across cultural contexts, using the Explorer to test scenarios and analyze alignment with diverse ethical frameworks.',
+    },
+    {
+      name: 'David Kim',
+      role: 'Business Owner',
+      description:
+        'David customizes AI interactions for his customers, using the Explorer to align responses with his brand’s values and ensure trustworthiness.',
+    },
+  ];
